requirements/test.txt
-r base.txt

pytest-django==4.1.0
factory-boy==3.1.0
Faker==5.0.2
pytest-cov==2.10.1
coveralls==2.2.0
freezegun==1.1.0
mock==4.0.3
